Fix sort panics and integrate trained NNUE weights
Major fixes:
- Fixed recurring sort comparison panics by implementing key-based sorting
  in order_moves_with_all_heuristics and all move ordering functions
- Added get_move_sort_key() to create deterministic sort keys for moves
- Replaced nested comparison functions with tuple-based sorting for guaranteed total order

NNUE integration:
- Added automatic loading of trained NNUE weights in ShogiEngine::new()
- Engine automatically loads nnue_weights_trained.json if available
- Added is_nnue_enabled() method to check NNUE status
- Added visible logging: 'info string Using NNUE evaluation' at search start
- Added [NNUE] indicator on first depth info line

Documentation updates:
- Updated NNUE_POST_TRAINING_NEXT_STEPS.md with continued training status
- Updated NNUE_INTEGRATION_COMPLETE.md with latest training info
- Updated NNUE_POST_TRAINING_GUIDE.md with automatic loading instructions
- Created nnue_integration_test.rs example

All sort operations now use deterministic key-based sorting to prevent
transitivity violations. NNUE weights are automatically loaded and
visible logging confirms when NNUE is active during searches.

+++ src/types/all.rs
@@ -5234,7 +5234,7 @@ pub struct TimeBudgetStats {
     /// Number of depths completed
     pub depths_completed: u8,
     /// Number of depths that exceeded budget
-    pub depths_exceeded_budget: u8,
+    pub depths_exceeded_budget: u32,
     /// Average time estimation accuracy (0.0 to 1.0)
     pub estimation_accuracy: f64,
 }
